@@ -1,5 +1,5 @@
-import * as bsv from 'bsv'
-import { BoostUtils } from '../boost-utils'
+import * as bsv from '../bsv'
+import { Utils } from '../utils'
 
 export class Digest32 {
   constructor(
@@ -8,13 +8,13 @@ export class Digest32 {
   }
 
   static fromHex(x: string): Digest32 {
-    return new Digest32(new Buffer(x, 'hex').reverse())
+    return new Digest32(Buffer.from(x, 'hex').reverse())
   }
 
   // reverse because of a horrible convention that exists in Bitcoin
   // which got started due to stupid a bug long ago.
   get hex(): string {
-    return new Buffer(this.data).reverse().toString('hex')
+    return Buffer.from(this.data).reverse().toString('hex')
   }
 
   get number(): bsv.crypto.BN {
@@ -30,6 +30,6 @@ export class Digest32 {
   }
 
   get utf8(): string {
-    return BoostUtils.trimBufferString(this.data, true)
+    return Utils.trimBufferString(this.data, true)
   }
 }
