@@ -1,5 +1,6 @@
-import * as bsv from 'bsv'
-import { BoostUtils } from '../boost-utils'
+import * as bsv from '../bsv'
+import { Utils } from '../utils'
+import { Digest32 } from './digest32'
 
 export class Bytes {
   constructor(
@@ -7,12 +8,16 @@ export class Bytes {
   ) {
   }
 
+  static fromHex(b: string): Bytes {
+    return new Bytes(Buffer.from(b, 'hex'))
+  }
+
   get buffer(): Buffer {
     return this.data
   }
 
   get hex(): string {
-    return new Buffer(this.buffer).toString('hex')
+    return this.buffer.toString('hex')
   }
 
   get string(): string {
@@ -20,6 +25,14 @@ export class Bytes {
   }
 
   get utf8(): string {
-    return BoostUtils.trimBufferString(this.buffer, true)
+    return Utils.trimBufferString(this.buffer, true)
+  }
+
+  get hash256(): Digest32 {
+    return new Digest32(bsv.crypto.Hash.sha256sha256(this.buffer))
+  }
+
+  get length(): number {
+    return this.data.length
   }
 }
