@@ -1,5 +1,4 @@
-import * as bsv from 'bsv'
-import { BoostUtils } from '../boost-utils'
+import { Utils } from '../utils'
 
 export class UInt16Little {
   constructor(
