@@ -1,10 +1,17 @@
 "use strict"
 var expect = require("chai").expect
-var index = require("../dist/index.js")
+var index = require("..")
 
-var bsv = require("bsv")
+var bsv = require("../lib/bsv")
 var sampleTx1 =
-  "010000000174d9f6dc235207fbbcdff7bdc412dcb375eb634da698ed164cc1e9aa1b88729a040000006b4830450221008596410738406e0e8589292a0e7a4d960e739025ab1859a3df6c77b4cf8c59ac0220733024f199162bc7b9ccd648aae56f0a0e307558a9827a26d35b1016de1865c54121025a77fe27d1db166b660205ff08b97f7dd87c7c68edaa2931895c2c8577f1a351ffffffff027d20000000000000e108626f6f7374706f777504000000002035b8fcb6882f93bddb928c9872198bcdf057ab93ed615ad938f24a63abde588104ffff001d14000000000000000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acb461590e000000001976a914ba6e459a2b505dc78e44e8c5874776c00890e16088ac00000000";
+  "010000000174d9f6dc235207fbbcdff7bdc412dcb375eb634da698ed164cc1e9aa1b88729a" +
+  "040000006b4830450221008596410738406e0e8589292a0e7a4d960e739025ab1859a3df6c" +
+  "77b4cf8c59ac0220733024f199162bc7b9ccd648aae56f0a0e307558a9827a26d35b1016de" +
+  "1865c54121025a77fe27d1db166b660205ff08b97f7dd87c7c68edaa2931895c2c8577f1a3" +
+  "51ffffffff027d20000000000000e108626f6f7374706f777504000000002035b8fcb6882f" +
+  "93bddb928c9872198bcdf057ab93ed615ad938f24a63abde588104ffff001d140000000000" +
+  "00000000000000000000000000000004000000002000000000000000000000000000000000" +
+  "000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acb461590e000000001976a914ba6e459a2b505dc78e44e8c5874776c00890e16088ac00000000";
 var sampleTx1Value = 8317
 var sampleTx1Vout = 0
 var sampleTx1id =
@@ -37,7 +44,20 @@ describe("boost #BoostPowJob create various getters and setters", () => {
   it("should decode multiple Boost outputs test_decode", async () => {
     const jobs = index.BoostPowJob.fromTransactionGetAllOutputs(
       new bsv.Transaction(
-        "01000000025ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda5456db5c050000006a473044022022a2f55f76d58508bc6185bdf53b1bbf2af3080c27117649db91e4bc8e31b080022031ed4ca57327e7880f3b207dc6330beaf9700f1d653390bff3e58764e95dcc3b41210225e89c2d8eab9166bb707d6601964d953cfe6c474445cd503b1397c25368e9bfffffffff5ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda5456db5c060000006b4830450221008a9614646eee30675bf89653dbe87ff6af100cea5e7dce26aad3d44098c576ab022073fb0676f3bff069b499ab772dc4cc288a2e1798b576b066284c29e54068171941210225e89c2d8eab9166bb707d6601964d953cfe6c474445cd503b1397c25368e9bfffffffff090000000000000000fdcc01006a2231394878696756345179427633744870515663554551797131707a5a56646f4175740e237269736b202366696e616e63650a746578742f706c61696e04746578741f7477657463685f7477746578745f313630313535323531393137312e747874017c223150755161374b36324d694b43747373534c4b79316b683536575755374d74555235035345540b7477646174615f6a736f6e046e756c6c0375726c046e756c6c07636f6d6d656e74046e756c6c076d625f7573657203353233057265706c79046e756c6c047479706504706f73740974696d657374616d70046e756c6c036170700674776574636807696e766f6963652464396564613934312d346666312d343336652d626362302d303035353238363861336161017c22313550636948473232534e4c514a584d6f53556157566937575371633768436676610d424954434f494e5f454344534122314a4a51654d4e5158326a6e724c68336d4148694555734c6b4a3476724d756433584c584834333861312f49376e37594e4d334f3656773845715956473867784c3666393751692f6a4e775233664f425a51567267666e41566b35626f4636386e4a6a627675394245614d744f61515146616631462b624c692f303dd5270000000000001976a91405186ff0710ed004229e644c0653b2985c648a2388ac6d040000000000001976a9147b935888ef076519e9c24aa71900b3761b6fd2bd88acbc020000000000002e2018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d0221e8825479a87c7f758875acb004000000000000e108626f6f7374706f777504000000002018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d04ffff001d146b7369722300000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acbc020000000000002e204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a932330221e8825479a87c7f758875acb004000000000000e108626f6f7374706f77750400000000204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a9323304ffff001d1465636e616e69662300000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac00000000"
+        "01000000025ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda54" +
+        "56db5c050000006a473044022022a2f55f76d58508bc6185bdf53b1bbf2af3080c27" +
+        "117649db91e4bc8e31b080022031ed4ca57327e7880f3b207dc6330beaf9700f1d65" +
+        "3390bff3e58764e95dcc3b41210225e89c2d8eab9166bb707d6601964d953cfe6c47" +
+        "4445cd503b1397c25368e9bfffffffff5ee4c00793c6081e30d90146a51e31d791a2" +
+        "71c95b478555f47e5cda5456db5c060000006b4830450221008a9614646eee30675b" +
+        "f89653dbe87ff6af100cea5e7dce26aad3d44098c576ab022073fb0676f3bff069b4" +
+        "99ab772dc4cc288a2e1798b576b066284c29e54068171941210225e89c2d8eab9166" +
+        "bb707d6601964d953cfe6c474445cd503b1397c25368e9bfffffffff090000000000" +
+        "000000fdcc01006a2231394878696756345179427633744870515663554551797131" +
+        "707a5a56646f4175740e237269736b202366696e616e63650a746578742f706c6169" +
+        "6e04746578741f7477657463685f7477746578745f31363031353532353139313731" +
+        "2e747874017c223150755161374b36324d694b43747373534c4b79316b6835365757" +
+        "55374d74555235035345540b7477646174615f6a736f6e046e756c6c0375726c046e" + "756c6c07636f6d6d656e74046e756c6c076d625f7573657203353233057265706c79046e756c6c047479706504706f73740974696d657374616d70046e756c6c036170700674776574636807696e766f6963652464396564613934312d346666312d343336652d626362302d303035353238363861336161017c22313550636948473232534e4c514a584d6f53556157566937575371633768436676610d424954434f494e5f454344534122314a4a51654d4e5158326a6e724c68336d4148694555734c6b4a3476724d756433584c584834333861312f49376e37594e4d334f3656773845715956473867784c3666393751692f6a4e775233664f425a51567267666e41566b35626f4636386e4a6a627675394245614d744f61515146616631462b624c692f303dd5270000000000001976a91405186ff0710ed004229e644c0653b2985c648a2388ac6d040000000000001976a9147b935888ef076519e9c24aa71900b3761b6fd2bd88acbc020000000000002e2018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d0221e8825479a87c7f758875acb004000000000000e108626f6f7374706f777504000000002018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d04ffff001d146b7369722300000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acbc020000000000002e204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a932330221e8825479a87c7f758875acb004000000000000e108626f6f7374706f77750400000000204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a9323304ffff001d1465636e616e69662300000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac00000000"
       )
     )
 
@@ -60,7 +80,12 @@ describe("boost #BoostPowJob create various getters and setters", () => {
 
     const jobObj = job.toASM();
     expect(jobObj).to.eql(
-      "626f6f7374706f77 OP_DROP 00000000 646c726f77206f6c6c6568000000000000000000000000000000000000000000 b3936a1a 0 00000000 0 OP_CAT OP_SWAP OP_5 OP_ROLL OP_DUP OP_TOALTSTACK OP_CAT OP_2 OP_PICK OP_TOALTSTACK OP_5 OP_ROLL OP_SIZE OP_4 OP_EQUALVERIFY OP_CAT OP_5 OP_ROLL OP_SIZE OP_8 OP_EQUALVERIFY OP_CAT OP_SWAP OP_CAT OP_HASH256 OP_SWAP OP_TOALTSTACK OP_CAT OP_CAT OP_SWAP OP_SIZE OP_4 OP_EQUALVERIFY OP_CAT OP_FROMALTSTACK OP_CAT OP_SWAP OP_SIZE OP_4 OP_EQUALVERIFY OP_CAT OP_HASH256 00 OP_CAT OP_BIN2NUM OP_FROMALTSTACK OP_SIZE OP_4 OP_EQUALVERIFY OP_3 OP_SPLIT OP_DUP OP_BIN2NUM OP_3 21 OP_WITHIN OP_VERIFY OP_TOALTSTACK OP_DUP OP_BIN2NUM 0 OP_GREATERTHAN OP_VERIFY 0000000000000000000000000000000000000000000000000000000000 OP_CAT OP_FROMALTSTACK OP_3 OP_SUB OP_8 OP_MUL OP_RSHIFT 00 OP_CAT OP_BIN2NUM OP_LESSTHAN OP_VERIFY OP_DUP OP_HASH160 OP_FROMALTSTACK OP_EQUALVERIFY OP_CHECKSIG"
+      "626f6f7374706f77 OP_DROP 00000000 " +
+      "646c726f77206f6c6c6568000000000000000000000000000000000000000000 " +
+      "b3936a1a 0 00000000 0 OP_CAT OP_SWAP OP_5 OP_ROLL OP_DUP OP_TOALTSTACK " +
+      "OP_CAT OP_2 OP_PICK OP_TOALTSTACK OP_5 OP_ROLL OP_SIZE OP_4 " +
+      "OP_EQUALVERIFY OP_CAT OP_5 OP_ROLL OP_SIZE OP_8 OP_EQUALVERIFY OP_CAT " +
+      "OP_SWAP OP_CAT OP_HASH256 OP_SWAP OP_TOALTSTACK OP_CAT OP_CAT OP_SWAP OP_SIZE OP_4 OP_EQUALVERIFY OP_CAT OP_FROMALTSTACK OP_CAT OP_SWAP OP_SIZE OP_4 OP_EQUALVERIFY OP_CAT OP_HASH256 00 OP_CAT OP_BIN2NUM OP_FROMALTSTACK OP_SIZE OP_4 OP_EQUALVERIFY OP_3 OP_SPLIT OP_DUP OP_BIN2NUM OP_3 21 OP_WITHIN OP_VERIFY OP_TOALTSTACK OP_DUP OP_BIN2NUM 0 OP_GREATERTHAN OP_VERIFY 0000000000000000000000000000000000000000000000000000000000 OP_CAT OP_FROMALTSTACK OP_3 OP_SUB OP_8 OP_MUL OP_RSHIFT 00 OP_CAT OP_BIN2NUM OP_LESSTHAN OP_VERIFY OP_DUP OP_HASH160 OP_FROMALTSTACK OP_EQUALVERIFY OP_CHECKSIG"
     )
   })
 
@@ -286,7 +311,15 @@ describe("boost #BoostPowJob create various getters and setters", () => {
 describe("boost #BoostPowString tryValidateJobProof", () => {
   it("tryValidateJobProof success sample test_decode", async () => {
     const job = index.BoostPowJob.fromRawTransaction(
-      "01000000025ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda5456db5c050000006a473044022022a2f55f76d58508bc6185bdf53b1bbf2af3080c27117649db91e4bc8e31b080022031ed4ca57327e7880f3b207dc6330beaf9700f1d653390bff3e58764e95dcc3b41210225e89c2d8eab9166bb707d6601964d953cfe6c474445cd503b1397c25368e9bfffffffff5ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda5456db5c060000006b4830450221008a9614646eee30675bf89653dbe87ff6af100cea5e7dce26aad3d44098c576ab022073fb0676f3bff069b499ab772dc4cc288a2e1798b576b066284c29e54068171941210225e89c2d8eab9166bb707d6601964d953cfe6c474445cd503b1397c25368e9bfffffffff090000000000000000fdcc01006a2231394878696756345179427633744870515663554551797131707a5a56646f4175740e237269736b202366696e616e63650a746578742f706c61696e04746578741f7477657463685f7477746578745f313630313535323531393137312e747874017c223150755161374b36324d694b43747373534c4b79316b683536575755374d74555235035345540b7477646174615f6a736f6e046e756c6c0375726c046e756c6c07636f6d6d656e74046e756c6c076d625f7573657203353233057265706c79046e756c6c047479706504706f73740974696d657374616d70046e756c6c036170700674776574636807696e766f6963652464396564613934312d346666312d343336652d626362302d303035353238363861336161017c22313550636948473232534e4c514a584d6f53556157566937575371633768436676610d424954434f494e5f454344534122314a4a51654d4e5158326a6e724c68336d4148694555734c6b4a3476724d756433584c584834333861312f49376e37594e4d334f3656773845715956473867784c3666393751692f6a4e775233664f425a51567267666e41566b35626f4636386e4a6a627675394245614d744f61515146616631462b624c692f303dd5270000000000001976a91405186ff0710ed004229e644c0653b2985c648a2388ac6d040000000000001976a9147b935888ef076519e9c24aa71900b3761b6fd2bd88acbc020000000000002e2018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d0221e8825479a87c7f758875acb004000000000000e108626f6f7374706f777504000000002018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d04ffff001d146b7369722300000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acbc020000000000002e204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a932330221e8825479a87c7f758875acb004000000000000e108626f6f7374706f77750400000000204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a9323304ffff001d1465636e616e69662300000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac00000000", 6)
+      "01000000025ee4c00793c6081e30d90146a51e31d791a271c95b478555f47e5cda5456" +
+      "db5c050000006a473044022022a2f55f76d58508bc6185bdf53b1bbf2af3080c271176" +
+      "49db91e4bc8e31b080022031ed4ca57327e7880f3b207dc6330beaf9700f1d653390bf" +
+      "f3e58764e95dcc3b41210225e89c2d8eab9166bb707d6601964d953cfe6c474445cd50" +
+      "3b1397c25368e9bfffffffff5ee4c00793c6081e30d90146a51e31d791a271c95b4785" +
+      "55f47e5cda5456db5c060000006b4830450221008a9614646eee30675bf89653dbe87f" +
+      "f6af100cea5e7dce26aad3d44098c576ab022073fb0676f3bff069b499ab772dc4cc28" +
+      "8a2e1798b576b066284c29e54068171941210225e89c2d8eab9166bb707d6601964d95" +
+      "3cfe6c474445cd503b1397c25368e9bfffffffff090000000000000000fdcc01006a22" + "31394878696756345179427633744870515663554551797131707a5a56646f4175740e237269736b202366696e616e63650a746578742f706c61696e04746578741f7477657463685f7477746578745f313630313535323531393137312e747874017c223150755161374b36324d694b43747373534c4b79316b683536575755374d74555235035345540b7477646174615f6a736f6e046e756c6c0375726c046e756c6c07636f6d6d656e74046e756c6c076d625f7573657203353233057265706c79046e756c6c047479706504706f73740974696d657374616d70046e756c6c036170700674776574636807696e766f6963652464396564613934312d346666312d343336652d626362302d303035353238363861336161017c22313550636948473232534e4c514a584d6f53556157566937575371633768436676610d424954434f494e5f454344534122314a4a51654d4e5158326a6e724c68336d4148694555734c6b4a3476724d756433584c584834333861312f49376e37594e4d334f3656773845715956473867784c3666393751692f6a4e775233664f425a51567267666e41566b35626f4636386e4a6a627675394245614d744f61515146616631462b624c692f303dd5270000000000001976a91405186ff0710ed004229e644c0653b2985c648a2388ac6d040000000000001976a9147b935888ef076519e9c24aa71900b3761b6fd2bd88acbc020000000000002e2018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d0221e8825479a87c7f758875acb004000000000000e108626f6f7374706f777504000000002018eaa89db7431a62d32510bd0253e0d9758d805ae629a3ca78245878cee60b1d04ffff001d146b7369722300000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acbc020000000000002e204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a932330221e8825479a87c7f758875acb004000000000000e108626f6f7374706f77750400000000204bd94af3e76f57ba13d790f02d5ff3ad7dd0451c2a31ecd2cb1fdbb0f3a9323304ffff001d1465636e616e69662300000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac6d1a0900000000001976a9147fb76fb504d39f83c6d86aa0763977a82b20eaf888ac00000000", 6)
 
     const jobProof = index.BoostPowJobProof.fromRawTransaction(
       "0100000001878bffb7d128225a2b2eaf98c1587a553bb88a4566ac81e5cb35075da613b76b060000009747304402200b6185b913fc852aa282dbd350e19ab8acbc96a6cf6c3fc0ccb4155a39c982c502202a792a1c3c831b41695136c0db0d5717eb501f8580dcbc6c1593de13276738eb412102d57db0c7046be47cb88e0e737b3037f7f29ef1c78ca412860eceffa617fee7ef043200098804b7768f5f08970200000000000004b008b2f5140af0c49c7e8c842243913bbea2491027cf4e6babffffffff01ab020000000000001976a9140af0c49c7e8c842243913bbea2491027cf4e6bab88ac00000000"
@@ -422,11 +455,23 @@ describe("boost #BoostPowJob createRedeemTransaction", () => {
     // const job =
     // index.BoostPowJob.fromRawTransaction('01000000013cdee5edfaec88f5ec5d4048c35ba1ed595a5c3dc8efc5360f8a26ec08621dcb010000006b483045022100af4682a0b78dc943f0f0f7fa85d3b4efe7291cad3f33a615e195f59b7d6c56f402207ee620e1848986128c95c07f1e2110fc1d165075bd6b4cbd2c1e24a9c566840b4121021e25de581fcd348717345e8f4c1996990b42f5914e1942b8356292100e43d427ffffffff02c922000000000000fd500108626f6f7374706f777504000000002035b8fcb6882f93bddb928c9872198bcdf057ab93ed615ad938f24a63abde588104ffff001d14000000000000000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e5279825488537f7653a269760120a1696b1d00000000000000000000000000000000000000000000000000000000007e6c5394986b557a8254887e557a8258887e7c7eaa517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c6b7e7e7c8254887e6c7e7c8254887eaa6c9f6976a96c88ace88df104000000001976a91432accdcb557ed57b9f01b4c42d69d4c9ea5d972a88ac00000000');
     const job = index.BoostPowJob.fromRawTransaction(
-      "010000000174d9f6dc235207fbbcdff7bdc412dcb375eb634da698ed164cc1e9aa1b88729a040000006b4830450221008596410738406e0e8589292a0e7a4d960e739025ab1859a3df6c77b4cf8c59ac0220733024f199162bc7b9ccd648aae56f0a0e307558a9827a26d35b1016de1865c54121025a77fe27d1db166b660205ff08b97f7dd87c7c68edaa2931895c2c8577f1a351ffffffff027d20000000000000e108626f6f7374706f777504000000002035b8fcb6882f93bddb928c9872198bcdf057ab93ed615ad938f24a63abde588104ffff001d14000000000000000000000000000000000000000004000000002000000000000000000000000000000000000000000000000000000000000000007e7c557a766b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c8254887eaa01007e816c825488537f7681530121a5696b768100a0691d00000000000000000000000000000000000000000000000000000000007e6c539458959901007e819f6976a96c88acb461590e000000001976a914ba6e459a2b505dc78e44e8c5874776c00890e16088ac00000000"
+      "010000000174d9f6dc235207fbbcdff7bdc412dcb375eb634da698ed164cc1e9aa1b88" +
+      "729a040000006b4830450221008596410738406e0e8589292a0e7a4d960e739025ab18" +
+      "59a3df6c77b4cf8c59ac0220733024f199162bc7b9ccd648aae56f0a0e307558a9827a" +
+      "26d35b1016de1865c54121025a77fe27d1db166b660205ff08b97f7dd87c7c68edaa29" +
+      "31895c2c8577f1a351ffffffff027d20000000000000e108626f6f7374706f77750400" +
+      "0000002035b8fcb6882f93bddb928c9872198bcdf057ab93ed615ad938f24a63abde58" +
+      "8104ffff001d1400000000000000000000000000000000000000000400000000200000" +
+      "0000000000000000000000000000000000000000000000000000000000007e7c557a76" +
+      "6b7e52796b557a8254887e557a8258887e7c7eaa7c6b7e7e7c8254887e6c7e7c825488" +
+      "7eaa01007e816c825488537f7681530121a5696b768100a0691d000000000000000000" +
+      "00000000000000000000000000000000000000007e6c539458959901007e819f6976a9" +
+      "6c88acb461590e000000001976a914ba6e459a2b505dc78e44e8c5874776c00890e16088ac00000000"
     )
     expect(job.txid).to.eql(
       "4eb545a588a21045495e74449b348ce1eb8f48ac95356c519a2a85a57731a518"
     )
+    expect(job.vout).to.eql(0)
     /*
     >>>>>>>>>>>>>>>>>>> checkShare >>>>>>>>>>>>>>>>>>>
     211193-Mar 15 05:43:40 ip-172-31-47-53 sserver[30281]: I0315 05:43:40.175833
@@ -614,6 +659,8 @@ describe("BoostPowJob", () => {
     const boostPowString = index.BoostPowString.fromString(
       "01000000646c726f77206f6c6c65480000000000000000000000000000000000000000002a96153663424ecfd483872e26e59bb02fd781a965df6575c437b0848e27d8aca6c8cb4dffff001dae5172dc"
     )
+
+    expect(boostPowString.hash).to.not.equal(undefined)
 
     expect(boostPowString.hash.hex).to.equal(
       "0000000086915e291fe43f10bdd8232f65e6eb64628bbb4d128be3836c21b6cc"
@@ -999,9 +1046,15 @@ describe("test serialize with empty values", () => {
     let job = index.BoostPowJob.fromObject({
       content: '00000000000000000000000058e3baf54d1f81304bafcf2a3e24db4bfd6b5cca',
       diff: 0.001
-    }).toScript()
+    })
 
-    expect(job).to.eql(index.BoostPowJob.fromScript(job).toScript())
+    let hex = job.toHex()
+    let buf = job.toBuffer()
+    let asm = job.toASM()
+
+    expect(hex).to.eql(index.BoostPowJob.fromHex(hex).toHex())
+    expect(buf).to.eql(index.BoostPowJob.fromBuffer(buf).toBuffer())
+    expect(asm).to.eql(index.BoostPowJob.fromASM(asm).toASM())
 
   })
 })
