@@ -1,6 +1,6 @@
 "use strict"
 const expect = require("chai").expect
-const index = require("../dist/index.js")
+const index = require("..")
 
 describe("boost #BoostPowString.fromObject", () => {
   it("should correctly decode from object to string", async () => {
@@ -20,7 +20,9 @@ describe("boost #BoostPowString.fromObject", () => {
     )
 
     expect(obj.toString()).to.eql(
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
+      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000" +
+      "006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8" +
+      "cb4db3936a1ae3143991"
     )
   })
 })
@@ -84,16 +86,9 @@ describe("boost #BoostPowString.fromString", () => {
   })
 
   it("should correctly decode Bitcoin header but invalid target", async () => {
-    try {
-      index.BoostPowString.fromString(
-        "020000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-      )
-    } catch (ex) {
-      expect(ex.message).to.equal("INVALID_POW")
-      return
-    }
-    // Should never get here
-    expect(true).to.equal(false)
+    expect(index.BoostPowString.fromString(
+      "020000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
+    ).valid()).to.equal(false)
   })
 
   it("should correctly decode from string to object", async () => {
@@ -115,16 +110,15 @@ describe("boost #BoostPowString.fromString", () => {
   })
 })
 
-describe("boost #BoostPowString validateProofOfWork ", () => {
+describe("boost #BoostPowString validateProofOfWork from string", () => {
   it("validProofOfWorkFromString success ", async () => {
-    const result = index.BoostPowString.validProofOfWorkFromString(
+    expect(index.BoostPowString.fromString(
       "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-    )
-    expect(result).to.eql(true)
+    ).valid()).to.eql(true)
   })
 
-  it("validProofOfWorkFromObject success ", async () => {
-    const result = index.BoostPowString.validProofOfWorkFromObject({
+  it("boost #BoostPowString validateProofOfWork from object", async () => {
+    expect(index.BoostPowString.fromObject({
       content:
         "0000000000000b60bc96a44724fd72daf9b92cf8ad00510b5224c6253ac40095",
       bits: 443192243,
@@ -133,23 +127,18 @@ describe("boost #BoostPowString validateProofOfWork ", () => {
       time: 1305200806,
       nonce: 2436437219,
       category: 1,
-    })
-
-    expect(result).to.eql(true)
+    }).valid()).to.eql(true)
   })
 
-  it("validProofOfWorkFromBuffer success ", async () => {
-    const buf = Buffer.from(
+  it("boost #BoostPowString validateProofOfWork from buffer", async () => {
+    expect(index.BoostPowString.fromBuffer(Buffer.from(
       "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991",
       "hex"
-    )
-
-    const result = index.BoostPowString.validProofOfWorkFromBuffer(buf)
-    expect(result).to.eql(true)
+    )).valid()).to.eql(true)
   })
 
-  it("validProofOfWorkFromObject success", async () => {
-    const result = index.BoostPowString.validProofOfWorkFromObject({
+  it("boost #BoostPowString validateProofOfWork from object, fail case", async () => {
+    expect(index.BoostPowString.fromObject({
       content:
         "0000000000000b60bc96a44724fd72daf9b92cf8ad00510b5224c6253ac40095",
       bits: 443192243,
@@ -158,9 +147,7 @@ describe("boost #BoostPowString validateProofOfWork ", () => {
       time: 1305200806,
       nonce: 1,
       category: 1,
-    })
-
-    expect(result).to.eql(false)
+    }).valid()).to.eql(false)
   })
 
   it("getDifficulty success", async () => {
@@ -169,39 +156,6 @@ describe("boost #BoostPowString validateProofOfWork ", () => {
     )
     const diff = boostPowString.difficulty
     expect(diff).to.eql(157416.40184364)
-  })
-
-  it("fromStringArray success ", async () => {
-    const result = index.BoostPowString.fromStringArray([
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991",
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991",
-    ])
-
-    const boostPowString1 = index.BoostPowString.fromString(
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-    )
-    const boostPowString2 = index.BoostPowString.fromString(
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-    )
-
-    expect(result).to.eql([boostPowString1, boostPowString2])
-  })
-
-  it("fromStringArray success ensure always sorted direction=desc", async () => {
-    const result = index.BoostPowString.fromStringArray([
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991",
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991",
-    ])
-
-    const boostPowString1 = index.BoostPowString.fromString(
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-    )
-    const boostPowString2 = index.BoostPowString.fromString(
-      "010000009500c43a25c624520b5100adf82cb9f9da72fd2447a496bc600b0000000000006cd862370395dedf1da2841ccda0fc489e3039de5f1ccddef0e834991a65600ea6c8cb4db3936a1ae3143991"
-    )
-
-    // todo
-    expect(result).to.eql([boostPowString1, boostPowString2])
   })
 })
 
@@ -442,6 +396,20 @@ describe("boost integration test ", () => {
     expect(userNonce.buffer).to.eql(userNonceBuffer)
   })
 
+  it("should get script version from locking script", async () => {
+    expect(jobBountyV1.scriptVersion).to.eql(1)
+    expect(jobBountyV1.useASICBoost).to.eql(false)
+
+    expect(jobBountyV2.scriptVersion).to.eql(2)
+    expect(jobBountyV2.useASICBoost).to.eql(true)
+
+    expect(jobContractV1.scriptVersion).to.eql(1)
+    expect(jobContractV1.useASICBoost).to.eql(false)
+
+    expect(jobContractV2.scriptVersion).to.eql(2)
+    expect(jobContractV2.useASICBoost).to.eql(true)
+  })
+
   const solutionBountyV1 = index.BoostPowJobProof.fromObject({
     signature: signatureHex,
     minerPubKeyHash: minerPubKeyHashHex,
@@ -582,7 +550,7 @@ describe("boost integration test ", () => {
     "OP_CAT OP_BIN2NUM OP_LESSTHAN OP_VERIFY OP_DUP OP_HASH160 " +
     "OP_FROMALTSTACK OP_EQUALVERIFY OP_CHECKSIG"
 
-const expectedLockingScriptBountyV2 =
+  const expectedLockingScriptBountyV2 =
     "626F6F7374706F77 OP_DROP D2040000 " +
     "68656C6C6F20616E696D616C0000000000000000000000000000000000000000 " +
     "D80F271E 74686973206973206120746167 C8010000 " +
